Separate most socket routes, the middleware, and fix some crashes
