+++ routes/login.js
@@ -5,8 +5,8 @@ const { logNumbers } = require("../modules/config")
 const { logger } = require("../modules/logger")
 const { Student } = require("../modules/student")
 const { STUDENT_PERMISSIONS, MANAGER_PERMISSIONS } = require("../modules/permissions")
+const { managerUpdate } = require("../modules/socketUpdates")
 const crypto = require('crypto')
-const { managerUpdate } = require("../sockets/init")
 
 module.exports = {
     run(app) {
