+++ sockets/help.js
@@ -1,8 +1,6 @@
-const { isLoggedIn } = require("../modules/authentication")
 const { classInformation } = require("../modules/class")
-const { logNumbers } = require("../modules/config")
-const { database } = require("../modules/database")
 const { logger } = require("../modules/logger")
+const { advancedEmitToClass } = require("../modules/socketUpdates")
 
 module.exports = {
     run(socket, socketUpdates) {
@@ -25,7 +23,23 @@ module.exports = {
 
                 logger.log('verbose', `[help] user=(${JSON.stringify(student)}`)
 
-                socketUpdates.updateClassPermissions()
+                socketUpdates.classPermissionUpdate()
+            } catch (err) {
+                logger.log('error', err.stack)
+            }
+        })
+
+        // Deletes help ticket
+        socket.on('deleteTicket', (student) => {
+            try {
+                logger.log('info', `[deleteTicket] ip=(${socket.handshake.address}) session=(${JSON.stringify(socket.request.session)})`)
+                logger.log('info', `[deleteTicket] student=(${student})`)
+
+                classInformation[socket.request.session.class].students[student].help = false
+
+                logger.log('verbose', `[deleteTicket] user=(${JSON.stringify(classInformation[socket.request.session.class].students[student])})`)
+
+                socketUpdates.classPermissionUpdate()
             } catch (err) {
                 logger.log('error', err.stack)
             }
