+++ sockets/middleware/rateLimiter.js
@@ -1,0 +1,51 @@
+const { logger } = require("../../modules/logger")
+const { rateLimits } = require("../../modules/socketUpdates")
+
+module.exports = {
+    order: 30,
+    run(socket, socketUpdates) {
+        // Rate limiter
+        socket.use(([event, ...args], next) => {
+            try {
+                const username = socket.request.session.username
+                const currentTime = Date.now()
+                const limit = 5
+                const timeFrame = 5000
+                const blockTime = 5000
+                const limitedRequests = ['pollResp', 'help', 'break']
+
+                logger.log('info', `[rate limiter] username=(${username}) currentTime=(${currentTime})`)
+                if (!rateLimits[username]) {
+                    rateLimits[username] = {}
+                }
+
+                const userRequests = rateLimits[username]
+                if (!limitedRequests.includes(event)) {
+                    next()
+                    return
+                }
+
+                userRequests[event] = userRequests[event] || []
+                userRequests[event] = userRequests[event].filter((timestamp) => currentTime - timestamp < timeFrame)
+                logger.log('verbose', `[rate limiter] userRequests=(${JSON.stringify(userRequests)})`)
+
+                if (userRequests[event].length >= limit) {
+                    socket.emit('message', `You are being rate limited. Please try again in a ${blockTime / 1000} seconds.`)
+                    next(new Error('Rate limited'))
+                    setTimeout(() => {
+                        try {
+                            userRequests[event].shift()
+                        } catch (err) {
+                            logger.log('error', err.stack);
+                        }
+                    }, blockTime)
+                } else {
+                    userRequests[event].push(currentTime)
+                    next()
+                }
+            } catch (err) {
+                logger.log('error', err.stack);
+            }
+        })        
+    }
+}
