+++ sockets/permChange.js
@@ -1,6 +1,4 @@
-const { isLoggedIn } = require("../modules/authentication")
 const { classInformation } = require("../modules/class")
-const { logNumbers } = require("../modules/config")
 const { database } = require("../modules/database")
 const { logger } = require("../modules/logger")
 const { io } = require("../modules/webServer")
