+++ sockets/pollCreation.js
@@ -1,77 +1,6 @@
 const { classInformation } = require("../modules/class")
 const { logger } = require("../modules/logger")
-
-// Functions
-// General functions
-function convertHSLToHex(hue, saturation, lightness) {
-	try {
-		logger.log('info', `[convertHSLToHex] hue=${hue}, saturation=${saturation}, lightness=${lightness}`)
-
-		// Normalize lightness to range 0-1
-		lightness /= 100;
-
-		// Calculate chroma
-		const chroma = saturation * Math.min(lightness, 1 - lightness) / 100;
-
-		// Function to get color component
-		function getColorComponent(colorIndex) {
-			try {
-				const colorPosition = (colorIndex + hue / 30) % 12;
-				const colorValue = lightness - chroma * Math.max(Math.min(colorPosition - 3, 9 - colorPosition, 1), -1);
-
-				// Return color component in hexadecimal format
-				return Math.round(255 * colorValue).toString(16).padStart(2, '0');
-			} catch (err) {
-				return err
-			}
-		}
-
-		// Return the hex color
-		logger.log('verbose', `[convertHSLToHex]  color=(${getColorComponent(0)}${getColorComponent(8)}${getColorComponent(4)})`)
-
-		let red = getColorComponent(0)
-		let green = getColorComponent(8)
-		let blue = getColorComponent(4)
-
-		if (red instanceof Error) throw red
-		if (green instanceof Error) throw green
-		if (blue instanceof Error) throw blue
-
-		return `#${red}${green}${blue}`;
-	} catch (err) {
-		return err
-	}
-}
-
-function generateColors(amount) {
-	try {
-		logger.log('info', `[generateColors] amount=(${amount})`)
-		// Initialize colors array
-		let colors = []
-
-		// Initialize hue
-		let hue = 0
-
-		// Generate colors
-		for (let i = 0; i < amount; i++) {
-			// Add color to the colors array
-			let color = convertHSLToHex(hue, 100, 50)
-
-			if (color instanceof Error) throw color
-
-			colors.push(color);
-
-			// Increment hue
-			hue += 360 / amount
-		}
-
-		// Return the colors array
-		logger.log('verbose', `[generateColors] colors=(${colors})`)
-		return colors
-	} catch (err) {
-		return err
-	}
-}
+const { generateColors } = require("../modules/util")
 
 module.exports = {
     run(socket, socketUpdates) {
@@ -151,6 +80,84 @@ module.exports = {
                 socketUpdates.virtualBarUpdate()
                 socketUpdates.classPermissionUpdate()
                 socket.emit('startPoll')
+            } catch (err) {
+                logger.log('error', err.stack);
+            }
+        })
+
+        socket.on('savePoll', (poll, id) => {
+            try {
+                logger.log('info', `[savePoll] ip=(${socket.handshake.address}) session=(${JSON.stringify(socket.request.session)})`)
+                logger.log('info', `[savePoll] poll=(${JSON.stringify(poll)}) id=(${id})`)
+
+                let userId = socket.request.session.userId
+
+                if (id) {
+                    database.get('SELECT * FROM custom_polls WHERE id=?', [id], (err, poll) => {
+                        try {
+                            if (err) throw err
+
+                            if (userId != poll.owner) {
+                                socket.emit('message', 'You do not have permission to edit this poll.')
+                                return
+                            }
+
+                            database.run('UPDATE custom_polls SET name=?, prompt=?, answers=?, textRes=?, blind=?, weight=?, public=? WHERE id=?', [
+                                poll.name,
+                                poll.prompt,
+                                JSON.stringify(poll.answers),
+                                poll.textRes,
+                                poll.blind,
+                                poll.weight,
+                                poll.public,
+                                id
+                            ], (err) => {
+                                try {
+                                    if (err) throw err
+
+                                    socket.emit('message', 'Poll saved successfully!')
+                                    socketUpdates.customPollUpdate(socket.request.session.username)
+                                } catch (err) {
+                                    logger.log('error', err.stack);
+                                }
+                            })
+                        } catch (err) {
+                            logger.log('error', err.stack);
+                        }
+                    })
+                } else {
+                    database.get('SELECT seq AS nextPollId from sqlite_sequence WHERE name = "custom_polls"', (err, nextPollId) => {
+                        try {
+                            if (err) throw err
+                            if (!nextPollId) logger.log('critical', '[savePoll] nextPollId not found')
+
+                            nextPollId = nextPollId.nextPollId + 1
+
+                            database.run('INSERT INTO custom_polls (owner, name, prompt, answers, textRes, blind, weight, public) VALUES (?, ?, ?, ?, ?, ?, ?, ?)', [
+                                userId,
+                                poll.name,
+                                poll.prompt,
+                                JSON.stringify(poll.answers),
+                                poll.textRes,
+                                poll.blind,
+                                poll.weight,
+                                poll.public
+                            ], (err) => {
+                                try {
+                                    if (err) throw err
+
+                                    classInformation[socket.request.session.class].students[socket.request.session.username].ownedPolls.push(nextPollId)
+                                    socket.emit('message', 'Poll saved successfully!')
+                                    socketUpdates.customPollUpdate(socket.request.session.username)
+                                } catch (err) {
+                                    logger.log('error', err.stack);
+                                }
+                            })
+                        } catch (err) {
+                            logger.log('error', err.stack);
+                        }
+                    })
+                }
             } catch (err) {
                 logger.log('error', err.stack);
             }
